Remove MiscSymFactory, moving functionality into SymFactory
No benefit to the split.
